Refactored db logic to make the codebase mcuh more cleaner
