+++ backend/.gitignore
@@ -9,3 +9,4 @@ wheels/
 # Virtual environments
 .venv
 .env
+database/test.db
